Log in and Registration screens
feat: Reusable views created for the header and button.

feat: Padding and white space adjusted on the login and registration screens.

feat: Replaced the button to link to the reusable button.

feat: Changed the login screen button colour to blue.

feat: Added header, form and button elements to the registration screen.

feat: Changed the header colour for the registration screen to green and the button colour to orange.